requirements.txt
couchbase==4.1.9
fastapi==0.104.1
httpx==0.25.1
pydantic==2.5.2
pydantic_settings==2.1.0
pytest==7.4.3
python-dotenv==1.0.0
uvicorn==0.24.0.post1
